refactor: split screenshot generator into focused modules
Extract types, constants, paths, overlays, and browser orchestration
into separate files under scripts/screenshots/ for clarity.
